sortIssues seperated to several cmps
