+++ src/components/issues/new-issue/NewIssue.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
+// TODO: compose
 import { compose } from 'redux';
 import styled from 'styled-components';
